Move SOKEmbedding to a separate file
